Move CanvasViewController mouse tests into its own file.
